adwords_tool/requirements.txt
google-ads==22.0.0
autohive_integrations_sdk
googleapis-common-protos 